pyproject.toml
[project]
name = "dailyme"
version = "0.1.0"
description = "Personalized AI news from newsletters — powered by OpenHands coding agents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.18",
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pgvector>=0.3.0",
    # Email ingestion
    "google-auth>=2.38.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.160.0",
    # Email sending
    "sendgrid>=6.11.0",
    # HTML parsing
    "beautifulsoup4>=4.12.0",
    "readability-lxml>=0.8.0",
    "markdownify>=0.14.0",
    "lxml>=5.0.0",
    # Embeddings (local, no API key)
    "sentence-transformers>=3.0.0",
    # Scheduling
    "apscheduler>=3.10.0",
    # Config
    "pydantic-settings>=2.7.0",
    # HTTP client
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.9.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.4",
]
